+++ src/main/scala/edu/ie3/simona/agent/grid/GridAgent.scala
@@ -16,7 +16,6 @@ import edu.ie3.simona.agent.state.AgentState.{Idle, Uninitialized}
 import edu.ie3.simona.agent.state.GridAgentState.{Initializing, SimulateGrid}
 import edu.ie3.simona.agent.{EnvironmentRefs, SimonaAgent}
 import edu.ie3.simona.config.SimonaConfig
-import edu.ie3.simona.exceptions.agent.GridAgentInitializationException
 import edu.ie3.simona.model.grid.GridModel
 import edu.ie3.simona.ontology.messages.PowerMessage.RequestGridPowerMessage
 import edu.ie3.simona.ontology.messages.SchedulerMessage.{
@@ -94,7 +93,6 @@ object GridAgent {
     */
   final case class FinishGridSimulationTrigger(tick: Long)
 }
-
 class GridAgent(
     val environmentRefs: EnvironmentRefs,
     simonaConfig: SimonaConfig,
@@ -157,7 +155,7 @@ class GridAgent(
           gridAgentInitData: GridAgentInitData
         ) =>
       // fail fast sanity checks
-      failFast(gridAgentInitData)
+      GridAgentFailFast.failFast(gridAgentInitData, simonaConfig)
 
       log.debug(
         s"Inferior Subnets: {}; Inferior Subnet Nodes: {}",
@@ -271,13 +269,4 @@ class GridAgent(
   // everything else
   whenUnhandled(myUnhandled())
 
-  private def failFast(gridAgentInitData: GridAgentInitData): Unit = {
-    if (
-      gridAgentInitData.superiorGridGates.isEmpty && gridAgentInitData.inferiorGridGates.isEmpty
-    )
-      throw new GridAgentInitializationException(
-        s"$actorName has neither superior nor inferior grids! This can either " +
-          s"be cause by wrong subnetGate information or invalid parametrization of the simulation!"
-      )
-  }
 }